(api) Make getting user a function, because it was repeating too much.

+++ backend/api/views.py
@@ -14,6 +14,18 @@ from django.http import JsonResponse
 
 logger = logging.getLogger(__name__)
 
+def getUser(request):
+    token_key = request.COOKIES.get('token')
+    if not token_key:
+        return Response({'detail': 'Authentication credentials were not provided.'}, status=status.HTTP_401_UNAUTHORIZED)
+    try:
+        token = Token.objects.get(key=token_key)
+        user = token.user
+        return user
+    except Token.DoesNotExist:
+        return Response({'detail': 'Invalid or expired token.'}, status=status.HTTP_401_UNAUTHORIZED)
+    
+
 class UserList(ListCreateAPIView):
     serializer_class = UserSerializer
 
@@ -160,36 +172,30 @@ def UpdateNote(request):
         return Response({'detail': 'Invalid or expired token.'}, status=status.HTTP_401_UNAUTHORIZED)
 
 @api_view(['POST'])
-@authentication_classes([SessionAuthentication, TokenAuthentication])
-@permission_classes([IsAuthenticated])
 def NewNote(request):
-    try:
-        user = request.user
-        try:
-            title = request.data['title']
-        except:
-            title = "New note"
-        try:
-            content = request.data['content']
-        except:
-            content = ""
-        serializer = NoteSerializer(data={"title": title, "content": content})
-        if not serializer.is_valid():
-            return Response(status=status.HTTP_400_BAD_REQUEST)
-        serializer.save(owner=user)
-        return Response(status=status.HTTP_200_OK)
-    except Exception as e:
-        if str(e) == "No Note matches the given query.":
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        logger.error(e)
-        return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+    user = getUser(request)
+    if isinstance(user, Response):
+        return user
+    
+    try: title = request.data['title']
+    except: title = "New note"
+
+    try: content = request.data['content']
+    except: content = ""
+
+    serializer = NoteSerializer(data={"title": title, "content": content})
+    if not serializer.is_valid():
+        return Response(status=status.HTTP_400_BAD_REQUEST)
+    serializer.save(owner=user)
+    return Response(status=status.HTTP_200_OK)
 
 @api_view(['POST'])
-@authentication_classes([SessionAuthentication, TokenAuthentication])
-@permission_classes([IsAuthenticated])
 def DeleteNote(request):
+    user = getUser(request)
+    if isinstance(user, Response):
+        return Response
     try:
-        user_id = request.user.id
+        user_id = user.id
         note_id = request.data["id"]
         note = get_object_or_404(Note, owner=user_id, id=note_id)
         note.delete()
